fix: Fix the file paths between the 2 js files
